Consolidate more modules (and break stuff)
