+++ opencog/cogserver/modules/commands/ModuleManagement.cc
@@ -1,5 +1,11 @@
-
-
+/*
+ * opencog/cogserver/modules/commands/ModuleManagement.cc
+ *
+ * Copyright (C) 2008 by OpenCog Foundation
+ * Copyright (C) 2022 Linas Vepstas
+ * Written by Gustavo Gama <gama@vettalabs.com>
+ * SPDX-License-Identifier: AGPL-3.0-or-later
+ */
 #include <sstream>
 #include <string>
 
@@ -22,9 +28,11 @@ public:                                                               \
 
 
 DEFINE_REQUEST(ListModulesRequest)
+DEFINE_REQUEST(LoadModuleRequest)
 
 using namespace opencog;
 
+// ====================================================================
 ListModulesRequest::ListModulesRequest(CogServer& cs) : Request(cs) {}
 ListModulesRequest::~ListModulesRequest() {}
 
@@ -47,3 +55,40 @@ bool ListModulesRequest::execute()
     return true;
 }
 
+// ====================================================================
+LoadModuleRequest::LoadModuleRequest(CogServer& cs) : Request(cs) {}
+LoadModuleRequest::~LoadModuleRequest() {}
+
+const RequestClassInfo&
+LoadModuleRequest::info()
+{
+    static const RequestClassInfo _cci(
+        "loadmodule",
+        "Load an opencog module",
+        "Usage: loadmodule <module>\n\n"
+        "Load the named opencog module"
+    );
+    return _cci;
+}
+
+bool LoadModuleRequest::execute()
+{
+    logger().debug("[LoadModuleRequest] execute");
+    std::ostringstream oss;
+    if (_parameters.empty()) {
+        oss << "invalid syntax: loadmodule <filename>" << std::endl;
+        send(oss.str());
+        return false;
+    }
+    std::string& filename = _parameters.front();
+
+    if (_cogserver.loadModule(filename)) {
+        oss << "done" << std::endl;
+        send(oss.str());
+        return true;
+    } else {
+        oss << "Unable to load module \"" << filename << "\". Check the server logs for details." << std::endl;
+        send(oss.str());
+        return false;
+    }
+}
